fixed: RandomGenerator 파일 분리
