+++ RacingCar/Controller/RacingGame.swift
@@ -8,27 +8,6 @@
 import Foundation
 
 
-protocol RandomGettable {
-    func random() -> Int
-    var randomRange: Range<Int> { get }
-}
-
-final class RandomGenerator: RandomGettable {
-    //MARK: - PROPERTIES
-    var randomRange: Range<Int>
-    
-    //MARK: - INITIALIZE
-    init (range: Range<Int> = 0..<10) {
-        self.randomRange = range
-    }
-    
-    //MARK: - METHOD
-    func random() -> Int {
-        return Int.random(in: randomRange)
-    }
-}
-
-
 final class RacingGame {
 
     //MARK: - DATA PROPERRIES
